prior_experiments/evals_share/requirements.txt
bs4==0.0.2
requests
google-generativeai
typing-extensions
openai 
